edge_computing/inference_service/requirements.txt
paho-mqtt
joblib
scikit-learn
pandas
influxdb-client
